backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic[email]
pydantic-settings
python-jose[cryptography]
pytest
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart
pytest-cov
httpx
flake8
mypy
types-python-jose
types-passlib
python-jose[cryptography]
python-dotenv
httpx